+++ lib/app.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+import 'package:rfwpg/screens/home_screen.dart';
+
+class App extends StatelessWidget {
+  const App({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'RFW PlayGround',
+      theme: ThemeData(
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+        useMaterial3: true,
+      ),
+      home: HomeScreen(),
+    );
+  }
+}
